backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
fastapi==0.128.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inngest==0.5.13
jcs==0.2.1
Jinja2==3.1.6
joblib==1.5.3
llama-index-core==0.14.12
llama-index-instrumentation==0.4.2
llama-index-readers-file==0.5.6
llama-index-workflows==2.12.2
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
ollama==0.6.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2026.1.15
requests==2.32.5
six==1.17.0
soupsieve==2.8.2
SQLAlchemy==2.0.45
starlette==0.50.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.0.1
yarl==1.22.0
